requirements_frontend.txt
# ── Local Frontend Requirements ───────────────────────────────────────────────
# Install with: pip install -r requirements_frontend.txt

torch>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
Flask>=3.0.0
